EC-199 require shading name field
This commit includes a temporary work around to two other test that
otherwise fail when the name field is required. This seems to be because
the way the user interaction seems to be mocked. When we add in a user
tab away from the field this seems to trigger a revalidation. It seems
that only after that has happened can the form be submitted successfully.