Configure Sentry to show more of the request body
This commit removes the "External Conditions" data from the request
body that we send to Sentry. It also sends the body as a string without
applying formatting - this means we save space and more of the request
body gets shown in Sentry (we were frequently running into Sentry's size
limit).