EC-927 Update heating systems summary to use '-'
This commit updates the summary page for heating systems to display '-'
instead of 'undefined' when a field has not yet been filled out. This is
necessitated by the auto-saving feature that allows a form to be saved
with partial data. The commit also displays more meaningful content for
the ecoDesignController and fuelType fields of the summary.