EC-955 Generate TS types from latest FHS schema
This commit brings in the latest FHS input schema as well as an OpenAPI
specification file that references it.

An new npm script is added & used to generate updated TypeScript types
from the OpenAPI definition file. This process does not work as
intended - likely due to `if-then-else` statements in the JSON schema.
Subsequent commits will update the schema manually to replace any
`if-then-else` blocks with schema composition logic (e.g. anyOf).