EC-955 Update HeatSources of hot water tanks
This commit updates the "HeatSource" property of Hot Water Tanks,
replacing the `if-then` blocks with schema composition logic. It also
adds a description to a nested field `heat_exchanger_surface_area` in
order to capture some logic that could not be replicated.

It also regenerates the TypeScript types from the updated schema, which
before this were not being generated correctly.