EC-955 Fix HeatSourceWet & update HeatBatteries
This commit fixes the "HeatSourceWet" property of the fhs JSON schema,
so that all properties are nested correctly within. It also updates the
HeatBattery type, replacing the `if-then` blocks with schema
composition logic.

It also regenerates the TypeScript types from the updated schema, which
before this were not being generated correctly for these sections.