EC-955 Update Building Elements in FHS schema
This commit updates the "BuildingElement" property in the FHS
schema, replacing `if-then` blocks with schema composition logic.
It also updates some existing schema composition logic so that types
can be generated correctly from it. This is done by bringing in the
"properties" field into the same block as the "required" field.

It also regenerates the TypeScript types from the updated schema, which
before this did not include all properties for this section.