EC-955 Update header tank & mains water properties
This commit adds properties to the header tank & mains water objects
of the "ColdWaterSource" section in the FHS schema. These were inferred
from example input files and had to be added in order to generate
correct TypeScript types for this section. Previously the generated type
was `Record<string, never>`.

It also regenerates the TypeScript types from the updated schema.