EC-955 Update SimulationTime object in FHS schema
This commit adds properties to the SimulationTime object in the FHS
schema. These were inferred from example input files and had to be added
in order to generate a correct TypeScript definition for this section.
Previous to this change the generated type was `Record<string, never>`.

It also regenerates the TypeScript types from the updated schema.